+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Fix: correct --atimes, --crtimes names 
 - Support --server, --sender internal arguments
 - Support --include, --filter arguments